+++ src/app/detalhes/[slug]/page.jsx
@@ -1,6 +1,7 @@
 'use client'
 import React, { useState } from 'react';
 import { getRestaurantById } from "@/lib/services/service"
+import ComentarioPaginaDetalhe from './comentarios'
 
 const HorariosFuncionamento = ({ horarios }) => {
   const horariosArray = Object.entries(horarios);
@@ -24,129 +25,6 @@ const HorariosFuncionamento = ({ horarios }) => {
           ))}
         </tbody>
       </table>
-    </div>
-  );
-};
-
-const Comentarios = () => {
-  const [comentario, setComentario] = useState('');
-  const [nomeUsuario, setNomeUsuario] = useState('');
-  const [avaliacao, setAvaliacao] = useState(0);
-  const [comentariosList, setComentariosList] = useState([]);
-  const [paginaAtual, setPaginaAtual] = useState(1);
-  const comentariosPorPagina = 3;
-
-  const handleComentarioChange = (e) => {
-    setComentario(e.target.value);
-  };
-
-  const handleNomeUsuarioChange = (e) => {
-    setNomeUsuario(e.target.value);
-  };
-
-  const handleAvaliacaoChange = (e) => {
-    setAvaliacao(parseInt(e.target.value, 10));
-  };
-
-  const handleComentarioSubmit = (e) => {
-    e.preventDefault();
-    if (nomeUsuario.trim() !== '' && comentario.trim() !== '' && avaliacao > 0 && avaliacao <= 5) {
-      const novoComentario = {
-        nomeUsuario,
-        avaliacao,
-        comentario,
-      };
-      setComentariosList([...comentariosList, novoComentario]);
-      setNomeUsuario('');
-      setAvaliacao(0);
-      setComentario('');
-    }
-  };
-
-  const indiceInicial = (paginaAtual - 1) * comentariosPorPagina;
-  const indiceFinal = paginaAtual * comentariosPorPagina;
-
-  const comentariosPaginaAtual = comentariosList.slice(indiceInicial, indiceFinal);
-
-  return (
-    <div className="comentarios-container container mt-4">
-      <h2>Comentários</h2>
-      <form onSubmit={handleComentarioSubmit} className="mb-4">
-        <div className="form-row">
-          <div className="col-md-6 mb-3">
-            <label htmlFor="nomeUsuario">Nome:</label>
-            <input
-              type="text"
-              id="nomeUsuario"
-              className="form-control"
-              value={nomeUsuario}
-              onChange={handleNomeUsuarioChange}
-            />
-          </div>
-          <div className="col-md-6 mb-3">
-            <label htmlFor="avaliacao">Avaliação:</label>
-            <select
-              id="avaliacao"
-              className="form-control"
-              value={avaliacao}
-              onChange={handleAvaliacaoChange}
-            >
-              <option value="0">Escolha uma avaliação</option>
-              {[1, 2, 3, 4, 5].map((numero) => (
-                <option key={numero} value={numero}>
-                  {numero}
-                </option>
-              ))}
-            </select>
-          </div>
-        </div>
-        <div className="form-group">
-          <label htmlFor="comentario">Comentário:</label>
-          <textarea
-            id="comentario"
-            className="form-control"
-            rows="4"
-            value={comentario}
-            onChange={handleComentarioChange}
-          ></textarea>
-        </div>
-        <button type="submit" className="btn btn-primary">
-          Enviar Comentário
-        </button>
-      </form>
-
-      <ul className="list-group mt-3">
-        {comentariosPaginaAtual.map((comentario, index) => (
-          <li key={index} className="list-group-item">
-            <p>
-              <strong>{comentario.nomeUsuario}</strong>
-              <br />
-              Avaliação: {comentario.avaliacao} estrelas
-            </p>
-            <p>{comentario.comentario}</p>
-          </li>
-        ))}
-      </ul>
-
-      <div className="pagination mt-3">
-        <button
-          className="btn btn-light"
-          onClick={() => setPaginaAtual(Math.max(1, paginaAtual - 1))}
-          disabled={paginaAtual === 1}
-        >
-          Anterior
-        </button>
-        <span className="mx-2">
-          Página {paginaAtual} de {Math.ceil(comentariosList.length / comentariosPorPagina)}
-        </span>
-        <button
-          className="btn btn-light"
-          onClick={() => setPaginaAtual(Math.min(Math.ceil(comentariosList.length / comentariosPorPagina), paginaAtual + 1))}
-          disabled={paginaAtual === Math.ceil(comentariosList.length / comentariosPorPagina)}
-        >
-          Próxima
-        </button>
-      </div>
     </div>
   );
 };
@@ -191,7 +69,7 @@ export default async function Detalhes({ params }) {
           ></iframe>
         </div>
       </div>
-      <Comentarios />
+      <ComentarioPaginaDetalhe idRestaurante={params.slug} />
     </>
   );
 };
